# Avoid copy of vectors in MakeRoute function